fix: 자료형 변경에 따른 코드 수정(#341)

+++ src/main/java/PodoeMarket/podoemarket/order/dto/response/OrderItemResponseDTO.java
@@ -22,6 +22,6 @@ public class OrderItemResponseDTO {
     private Long scriptPrice;
     private Integer performanceAmount;
     private Long performancePrice;
-    private Integer performanceTotalPrice;
+    private Long performanceTotalPrice;
     private Long totalPrice;
 }
